+++ js/main.js
@@ -1,63 +1,2 @@
-const comments = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце-концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как-будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const descriptions = [
-  'Сколько нежности и любви в этом простом и добром кадре',
-  'Давайте внимательнее рассмотрим изображение',
-  'Затусили с друзьями на море',
-  'Отдыхаем...',
-  'Цените каждое мгновенье. Цените тех, кто рядом с вами и отгоняйте все сомненья.',
-  'Вот это тачка!',
-  'Отличное кафе.Как же круто тут кормят',
-  'Господи, это такая милота.',
-  'Хорошо, когда в жизни есть верные друзья ',
-  'Прикольно',
-];
-const names = ['Николай', 'Максим', 'Андрей', 'Алексей', 'Тимур', 'Степан', 'Ольга', 'Наталья', 'Вероника', 'Валентина', 'Дмитрий', 'Олег', 'Сергей'];
-
-const getRandomPositiveInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-
-const getRandomArrayElement = (array) =>
-  array[getRandomPositiveInteger(0, array.length - 1)];
-
-const createMessage = () =>
-  Array.from({ length: getRandomPositiveInteger(1, 2) }, () =>
-    getRandomArrayElement(comments)
-  ).join(' ');
-
-const createComment = (index) => ({
-  id: index,
-  avatar: `img/avatar-${getRandomPositiveInteger(1, 6)}.svg`,
-  message: createMessage(),
-  name: getRandomArrayElement(names),
-});
-
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(descriptions),
-  likes: getRandomPositiveInteger(15, 200),
-  comments: Array.from(
-    { length: getRandomPositiveInteger(0, 6) },
-    (_, commentIndex) => createComment(commentIndex + 1)
-  ),
-});
-
-const getPictures = () =>
-  Array.from({ length: 25 }, (_, pictureIndex) =>
-    createPicture(pictureIndex + 1)
-  );
-
-(getPictures());
+import {getPictures} from './data.js';
+getPictures();
